requirements_order.txt
spacy
pathlib
logging
pyyaml
transformers
ultralytics
nudenet
deepface
pytesseract
deepdanbooru
tensorflow==2.11.0 
keras==2.11.0
datetime
tensorflow_io
clip_interrogator
imagehash
numpy==1.24.4












